requirements.txt
Django==1.9.13
djangorestframework==3.4.7
gunicorn==19.5.0
raven==5.19.0
requests==2.10.0
whitenoise==3.1
git+https://github.com/uktrade/directory-sso-api-client.git@1.0.1#egg=directory_sso_api_client
git+https://github.com/uktrade/directory-constants.git@2.0.0#egg=directory_constants
git+https://github.com/uktrade/directory-signature-auth.git@v0.1.1#egg=sigauth
git+https://github.com/uktrade/directory-api-client-external.git@0.2.0#egg=directory_api_external
git+https://github.com/uktrade/directory-header-footer.git@v0.2.0#egg=directory_header_footer